samples/ai/chat-demo/requirements.txt
Flask[async]==3.0.0
flask-sock==0.7.0
openai>=1.0.0
websockets>=11.0.0
flask-cors>=4.0.0
azure-storage-blob>=12.14.0
azure-messaging-webpubsubservice>=1.0.0
azure-identity>=1.15.0
asgiref>=3.7.2
python-dotenv>=1.0.0
azure-data-tables>=12.5.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
